Various fixes and updates. Confirmed working for TLM

+++ Decom/ByteManipulation.cpp
@@ -86,7 +86,7 @@ namespace ByteManipulation {
 		return std::stoul(s_result, nullptr, 2);
 	}
 
-	uint16_t mergeBytes16(uint8_t& initialByte, uint8_t& extraByte1)
+	uint32_t mergeBytes16(uint8_t& initialByte, uint8_t& extraByte1)
 	{
 		std::string b1 = std::bitset<8>(initialByte).to_string();
 		std::string b2 = std::bitset<8>(extraByte1).to_string();
